+++ src/org/elixir_lang/structure_view/element/EExFunctionFromHead.kt
@@ -1,0 +1,21 @@
+package org.elixir_lang.structure_view.element
+
+import com.intellij.ide.util.treeView.smartTree.TreeElement
+import com.intellij.navigation.ItemPresentation
+import org.elixir_lang.Visibility
+import org.elixir_lang.navigation.item_presentation.NameArity
+import org.elixir_lang.navigation.item_presentation.Parent
+import org.elixir_lang.psi.call.Call
+import org.elixir_lang.structure_view.element.modular.Modular
+
+class EExFunctionFromHead(private val eexFunctionFrom: EExFunctionFrom) : Element<Call>(eexFunctionFrom.call), Presentable, Visible {
+    override fun getPresentation(): ItemPresentation =
+            org.elixir_lang.navigation.item_presentation.CallDefinitionHead(
+                    eexFunctionFrom.presentation as NameArity,
+                    eexFunctionFrom.visibility()!!,
+                    navigationItem
+            )
+
+    override fun getChildren(): Array<TreeElement> = emptyArray()
+    override fun visibility(): Visibility? = eexFunctionFrom.visibility()
+}