+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/exception.ex
@@ -1,0 +1,7 @@
+defmodule MyException do
+  defexception [:message]
+
+  def extra do
+    e<caret>
+  end
+end