+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/to.ex
@@ -1,0 +1,5 @@
+defmodule To do
+  def source do
+
+  end
+end